// src/app/core/api/generated/models/order-detail-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { OrderItemDetailDto } from '../models/order-item-detail-dto';
export interface OrderDetailDto {
  created_at: string;
  id: string;
  items: Array<OrderItemDetailDto>;
  shipping_address: string;
  status: string;
  total: (number | string);
  updated_at: string;
  user_id: string;
}